requirements.txt
wikipedia-api==0.6.0
requests==2.31.0
Pillow==10.2.0
moviepy==1.0.3
edge-tts==6.1.9
numpy==1.26.3

